requirements.txt
dlib==19.24.1
Flask==2.2.3
imutils==0.5.4
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.5.0
torch==2.0.0
torchvision==0.15.1
Werkzeug==2.2.3
